requirements.txt
fastapi[all]
fastapi_cache2
pydantic_settings
uvicorn
requests
httpx
bs4
aiohttp==3.9.1
aiohttp-client-cache==0.8.2
aiosqlite
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
matplotlib
numpy
statistics

